requirements.txt
asgiref==3.7.2
Django==4.1
django-ckeditor==6.7.0
django-js-asset==2.1.0
Pillow==10.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
